Updated readme with some feature showcase + other changes
Changes:
- Added example project with code snippets
- Updated README
- Updated for changes in Mediator package
- Some small bug fixes

+++ src/Sels.Hivemind/Sels.HiveMind.Colony/Components/EventHandlers/DeletionDaemonAutoCreator.cs
@@ -37,7 +37,7 @@ namespace Sels.HiveMind.Colony.EventHandlers
         protected DeletionDaemonAutoCreator() { }
 
         /// <inheritdoc/>
-        public ushort? Priority => null;
+        public byte? Priority => null;
 
         /// <inheritdoc/>
         public virtual Task HandleAsync(IEventListenerContext context, ColonyCreatedEvent @event, CancellationToken token)
