+++ src/Sels.Hivemind/Sels.HiveMind.Colony/Components/EventHandlers/LockMonitorAutoCreator.cs
@@ -35,7 +35,7 @@ namespace Sels.HiveMind.Colony.EventHandlers
         protected LockMonitorAutoCreator() { }
 
         /// <inheritdoc/>
-        public ushort? Priority => null;
+        public byte? Priority => null;
 
         /// <inheritdoc/>
         public virtual Task HandleAsync(IEventListenerContext context, ColonyCreatedEvent @event, CancellationToken token)
