+++ src/Sels.Hivemind/Sels.HiveMind.Core/Components/EventHandlers/BackgroundJobCleanupTrigger.cs
@@ -30,7 +30,7 @@ namespace Sels.HiveMind.EventHandlers
 
         // Properties
         /// <inheritdoc/>
-        public ushort? Priority => null; // Always go last just in case other handlers throw exceptions
+        public byte? Priority => null; // Always go last just in case other handlers throw exceptions
 
         /// <inheritdoc cref="BackgroundJobProcessTrigger"/>
         /// <param name="options">Used to access the HiveMind options for completed job retention</param>
